+++ Packages/UGF.Logs/Editor/LogEditorSettingsProviderEditor.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;
@@ -6,67 +7,116 @@ namespace UGF.Logs.Editor
 {
     internal class LogEditorSettingsProviderEditor : SettingsProvider
     {
-        private LogEditorSettings m_settings;
-        private readonly GUIContent m_contentApplyAll = new GUIContent("Apply All", "Apply to all Build Target Groups");
+        private readonly string m_groupPrefKey = "UGF.Logs.Editor.LogEditorSettingsProviderEditor.Group";
+        private readonly List<GroupItem> m_groups = new List<GroupItem>();
+        private int m_group;
+        private SerializedObject m_serializedObject;
+        private Styles m_styles;
+        private Content m_content = new Content();
+
+        private class GroupItem
+        {
+            public BuildTargetGroup BuildTargetGroup { get; set; }
+            public GUIContent ContentName { get; set; }
+            public bool GroupAll { get; set; }
+        }
+
+        private class Data : ScriptableObject
+        {
+            [SerializeField] private LogEditorSettings m_settings = new LogEditorSettings();
+
+            public LogEditorSettings Settings { get { return m_settings; } set { m_settings = value; } }
+        }
+
+        private class Styles
+        {
+            public GUIStyle DropdownButton { get; } = "DropDownButton";
+        }
+
+        private class Content
+        {
+            public GUIContent LogInfoLabel { get; } = new GUIContent("Info", "Determines whether 'UGF_LOG_INFO' define is specified for selected platform.");
+            public GUIContent LogDebugLabel { get; } = new GUIContent("Debug", "Determines whether 'UGF_LOG_DEBUG' define is specified for selected platform.");
+            public GUIContent LogWarningLabel { get; } = new GUIContent("Warning", "Determines whether 'UGF_LOG_WARNING' define is specified for selected platform.");
+            public GUIContent LogErrorLabel { get; } = new GUIContent("Error", "Determines whether 'UGF_LOG_ERROR' define is specified for selected platform.");
+            public GUIContent LogExceptionLabel { get; } = new GUIContent("Exception", "Determines whether 'UGF_LOG_EXCEPTION' define is specified for selected platform.");
+        }
 
         public LogEditorSettingsProviderEditor() : base("Project/UGF/Logs", SettingsScope.Project, new[] { "Log" })
         {
+            var groupAll = new GroupItem
+            {
+                ContentName = new GUIContent("All Platforms"),
+                GroupAll = true
+            };
+
+            m_groups.Add(groupAll);
+
+            foreach (BuildTargetGroup buildTargetGroup in LogEditorUtility.InternalGetAllBuildTargetGroupValues())
+            {
+                var group = new GroupItem
+                {
+                    BuildTargetGroup = buildTargetGroup,
+                    ContentName = new GUIContent(buildTargetGroup.ToString())
+                };
+
+                m_groups.Add(group);
+            }
         }
 
         public override void OnActivate(string searchContext, VisualElement rootElement)
         {
-            m_settings = LogEditorUtility.GetSettings();
+            Load();
         }
 
         public override void OnDeactivate()
         {
-            if (m_settings != null && m_settings.Changed)
+            if (m_serializedObject != null)
             {
-                if (EditorUtility.DisplayDialog("Unapplied settings", "Unapplied settings for 'Logs'.", "Apply", "Revert"))
-                {
-                    Apply();
-                }
-                else
-                {
-                    Revert();
-                }
+                Save();
             }
         }
 
         public override void OnGUI(string searchContext)
         {
+            if (m_styles == null)
+            {
+                m_styles = new Styles();
+            }
+
+            m_serializedObject.UpdateIfRequiredOrScript();
+
+            GroupItem group = m_groups[m_group];
+
+            SerializedProperty propertyInfo = m_serializedObject.FindProperty("m_settings.m_logInfo");
+            SerializedProperty propertyDebug = m_serializedObject.FindProperty("m_settings.m_logDebug");
+            SerializedProperty propertyWarning = m_serializedObject.FindProperty("m_settings.m_logWarning");
+            SerializedProperty propertyError = m_serializedObject.FindProperty("m_settings.m_logError");
+            SerializedProperty propertyException = m_serializedObject.FindProperty("m_settings.m_logException");
+
             using (new EditorGUI.IndentLevelScope())
             {
-                m_settings.LogInfo = EditorGUILayout.Toggle("Info", m_settings.LogInfo);
-                m_settings.LogDebug = EditorGUILayout.Toggle("Debug", m_settings.LogDebug);
-                m_settings.LogWarning = EditorGUILayout.Toggle("Warning", m_settings.LogWarning);
-                m_settings.LogError = EditorGUILayout.Toggle("Error", m_settings.LogError);
-                m_settings.LogException = EditorGUILayout.Toggle("Exception", m_settings.LogException);
+                EditorGUILayout.PropertyField(propertyInfo, m_content.LogInfoLabel);
+                EditorGUILayout.PropertyField(propertyDebug, m_content.LogDebugLabel);
+                EditorGUILayout.PropertyField(propertyWarning, m_content.LogWarningLabel);
+                EditorGUILayout.PropertyField(propertyError, m_content.LogErrorLabel);
+                EditorGUILayout.PropertyField(propertyException, m_content.LogExceptionLabel);
 
-                using (new EditorGUI.DisabledScope(!m_settings.Changed))
                 using (new EditorGUILayout.HorizontalScope())
                 {
                     GUILayout.FlexibleSpace();
 
-                    if (GUILayout.Button("Revert"))
-                    {
-                        Revert();
-                    }
+                    Rect rect = GUILayoutUtility.GetRect(group.ContentName, m_styles.DropdownButton, GUILayout.MinWidth(100F));
 
-                    if (GUILayout.Button("Apply"))
+                    if (EditorGUI.DropdownButton(rect, group.ContentName, FocusType.Keyboard, m_styles.DropdownButton))
                     {
-                        Apply();
-                    }
-
-                    if (GUILayout.Button(m_contentApplyAll))
-                    {
-                        ApplyAll();
+                        ShowGroupMenu(rect);
                     }
 
                     EditorGUILayout.Space();
                 }
 
-                if (m_settings.Changed)
+                if (group.GroupAll || group.BuildTargetGroup == EditorUserBuildSettings.selectedBuildTargetGroup)
                 {
                     EditorGUILayout.Space();
                     EditorGUILayout.HelpBox("Applying settings will trigger project recompilation.", MessageType.Info);
@@ -78,25 +128,100 @@ namespace UGF.Logs.Editor
                     EditorGUILayout.HelpBox("Development Build is active, all setting are ignored and all logs included in build.", MessageType.Info);
                 }
             }
+
+            if (m_serializedObject.hasModifiedProperties)
+            {
+                m_serializedObject.ApplyModifiedProperties();
+
+                Save();
+            }
+
+            m_serializedObject.ApplyModifiedProperties();
         }
 
-        private void Revert()
+        private void SetGroup(int index)
         {
-            m_settings.Revert();
+            if (m_group != index)
+            {
+                Save();
+
+                m_group = index;
+
+                EditorPrefs.SetInt(m_groupPrefKey, m_group);
+
+                Load();
+                Repaint();
+            }
         }
 
-        private void Apply()
+        private void Load()
         {
-            m_settings.Apply();
+            m_group = EditorPrefs.GetInt(m_groupPrefKey);
 
-            LogEditorUtility.SetSettings(m_settings);
+            GroupItem group = m_groups[m_group];
+
+            m_serializedObject = group.GroupAll ? LoadDataAll() : LoadData(group.BuildTargetGroup);
         }
 
-        private void ApplyAll()
+        private void Save()
         {
-            m_settings.Apply();
+            GroupItem group = m_groups[m_group];
+            var data = (Data)m_serializedObject.targetObject;
 
-            LogEditorUtility.SetSettingsAll(m_settings);
+            if (group.GroupAll)
+            {
+                LogEditorUtility.SetSettingsAll(data.Settings);
+            }
+            else
+            {
+                LogEditorUtility.SetSettings(data.Settings, group.BuildTargetGroup);
+            }
+        }
+
+        private SerializedObject LoadDataAll()
+        {
+            var targets = new List<Object>();
+
+            foreach (BuildTargetGroup buildTargetGroup in LogEditorUtility.InternalGetAllBuildTargetGroupValues())
+            {
+                var data = ScriptableObject.CreateInstance<Data>();
+
+                data.Settings = LogEditorUtility.GetSettings(buildTargetGroup);
+
+                targets.Add(data);
+            }
+
+            return new SerializedObject(targets.ToArray());
+        }
+
+        private SerializedObject LoadData(BuildTargetGroup buildTargetGroup)
+        {
+            var data = ScriptableObject.CreateInstance<Data>();
+
+            data.Settings = LogEditorUtility.GetSettings(buildTargetGroup);
+
+            return new SerializedObject(data);
+        }
+
+        private void ShowGroupMenu(Rect position)
+        {
+            var menu = new GenericMenu();
+
+            for (int i = 0; i < m_groups.Count; i++)
+            {
+                GroupItem group = m_groups[i];
+
+                menu.AddItem(group.ContentName, i == m_group, OnGroupMenuItemSelected, i);
+            }
+
+            menu.DropDown(position);
+        }
+
+        private void OnGroupMenuItemSelected(object userData)
+        {
+            int index = (int)userData;
+
+            SetGroup(index);
         }
     }
 }